added the params module

+++ construct_R.f90
@@ -13,11 +13,11 @@
 ! 	R	 
 SUBROUTINE construct_R(R,theta,snap_pos,D,nsnap,Pc)
    IMPLICIT NONE
-   REAL :: rxy
+   DOUBLE PRECISION :: rxy
    INTEGER :: D, Nsnap, Pc
    INTEGER :: ii, jj
-   REAL :: R(nsnap,nsnap),theta(D)
-   REAL :: snap_pos(nsnap,D)
+   DOUBLE PRECISION :: R(nsnap,nsnap),theta(D)
+   DOUBLE PRECISION :: snap_pos(nsnap,D)
    DO ii=1,nsnap
       DO jj=ii,nsnap
          R(ii,jj) = rxy(theta,snap_pos(ii,:),snap_pos(jj,:),D,Pc)
